refactor: Precompile markdown
