+++ plugins/precompile-markdown/gh-emoji/index.js
