+++ src/assets/_headers
@@ -4,9 +4,6 @@
     Link:""
 /*.css
     Link:""
-/*.md
-    Link:""
-	Content-Type: text/x-markdown
 /*.ico
     Link:""
 /*.svg
